Set up the basic scaffolding for cachier.
This sets up a basic framework for stamping out a controller that acts on a duck type, which can be fed several different GroupVersionResource tuples to elicit different behaviors.

For the purposed of example, I have used a pre-existing Duck type from knative/pkg, but that will change in a subsequent commit.